feat: move non-retry to exceptions
