+++ sdks/python/hatchet_sdk/exceptions.py
@@ -1,0 +1,2 @@
+class NonRetryableException(Exception):
+    pass
